requirements.txt
absl-py==0.8.1
astor==0.8.0
audioread==2.1.8
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
crepe==0.0.9
cycler==0.10.0
dask==1.2.2
distributed==1.28.1
fuzzywuzzy==0.17.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
HeapDict==1.0.1
hmmlearn==0.2.2
idna==2.8
imageio==2.6.1
jmetalpy==1.5.1
joblib==0.14.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.30.0
locket==0.2.0
Markdown==3.1.1
matplotlib==3.1.1
msgpack==0.6.2
numba==0.46.0
numpy==1.17.3
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==0.25.3
partd==1.0.0
patsy==0.5.1
Pillow==6.2.1
plotly==4.3.0
protobuf==3.10.0
psutil==5.6.5
py4j==0.10.7
pyacoustid==1.1.7
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydub==0.23.1
pyparsing==2.4.4
pyspark==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.13.0
sklearn==0.0
sortedcontainers==2.1.0
statsmodels==0.10.1
tblib==1.5.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
toolz==0.10.0
tornado==6.0.3
tqdm==4.38.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
zict==1.0.0
